update login flow, move login logic to userService
